examples/python/requirements.txt
# Python dependencies for JWT examples
pyjwt==2.8.0
cryptography==41.0.7

# Flask example dependencies (optional)
flask==3.0.0

# FastAPI example dependencies (optional)
fastapi==0.104.1
uvicorn[standard]==0.24.0

